requirements.txt
python-telegram-bot>=21.0.0
fastapi>=0.115.0
uvicorn>=0.30.0
requests>=2.31.0
python-dotenv>=1.0.1
SQLAlchemy>=2.0.0
aiosqlite>=0.20.0
APScheduler>=3.10.4
googletrans==4.0.0rc1